Fix bug: call alter config for all the generated configs, and reorganize packages
