+++ pkg/resources/kafka/configmap_test.go
@@ -15,14 +15,18 @@
 package kafka
 
 import (
+	"reflect"
+	"sort"
 	"testing"
 
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/scheme"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/banzaicloud/kafka-operator/api/v1beta1"
 	"github.com/banzaicloud/kafka-operator/pkg/resources"
+	"github.com/banzaicloud/kafka-operator/pkg/util"
 )
 
 func TestGenerateBrokerConfig(t *testing.T) {
@@ -239,5 +243,170 @@ zookeeper.connect=example.zk:2181/`,
 				t.Errorf("the expected config is %s, received: %s", test.expectedConfig, generatedConfig)
 			}
 		})
+	}
+}
+
+func TestShouldRefreshOnlyPerBrokerConfigs(t *testing.T) {
+	testCases := []struct {
+		Description    string
+		CurrentConfigs map[string]string
+		DesiredConfigs map[string]string
+		Result         bool
+	}{
+		{
+			Description: "configs did not change",
+			CurrentConfigs: map[string]string{
+				"unmodified_config_1": "unmodified_value_1",
+				"unmodified_config_2": "unmodified_value_2",
+			},
+			DesiredConfigs: map[string]string{
+				"unmodified_config_1": "unmodified_value_1",
+				"unmodified_config_2": "unmodified_value_2",
+			},
+			Result: true,
+		},
+		{
+			Description: "only non per-broker config changed",
+			CurrentConfigs: map[string]string{
+				"unmodified_config_1": "unmodified_value_1",
+				"unmodified_config_2": "modified_value_2",
+			},
+			DesiredConfigs: map[string]string{
+				"unmodified_config_1": "unmodified_value_1",
+				"unmodified_config_2": "modified_value_3",
+			},
+			Result: false,
+		},
+		{
+			Description: "only per-broker config changed",
+			CurrentConfigs: map[string]string{
+				"unmodified_config_1": "unmodified_value_1",
+				"ssl.client.auth":     "modified_value_2",
+			},
+			DesiredConfigs: map[string]string{
+				"unmodified_config_1": "unmodified_value_1",
+				"ssl.client.auth":     "modified_value_3",
+			},
+			Result: true,
+		},
+		{
+			Description: "per-broker and non per-broker configs have changed",
+			CurrentConfigs: map[string]string{
+				"modified_config_1": "modified_value_1",
+				"ssl.client.auth":   "modified_value_3",
+			},
+			DesiredConfigs: map[string]string{
+				"modified_config_1": "modified_value_2",
+				"ssl.client.auth":   "modified_value_4",
+			},
+			Result: false,
+		},
+		{
+			Description: "security protocol map can be changed as a per-broker config",
+			CurrentConfigs: map[string]string{
+				"listener.security.protocol.map": "listener1:protocol1,listener2:protocol2",
+			},
+			DesiredConfigs: map[string]string{
+				"listener.security.protocol.map": "listener1:protocol1,listener3:protocol3",
+			},
+			Result: true,
+		},
+		{
+			Description: "security protocol map can't be changed as a per-broker config",
+			CurrentConfigs: map[string]string{
+				"listener.security.protocol.map": "listener1:protocol1,listener2:protocol2",
+			},
+			DesiredConfigs: map[string]string{
+				"listener.security.protocol.map": "listener1:protocol1,listener2:protocol3",
+			},
+			Result: false,
+		},
+		{
+			Description:    "security protocol map added as config",
+			CurrentConfigs: map[string]string{},
+			DesiredConfigs: map[string]string{
+				"listener.security.protocol.map": "listener1:protocol1,listener2:protocol2",
+			},
+			Result: true,
+		},
+	}
+	logger := util.CreateLogger(false, false)
+	for i, testCase := range testCases {
+		if shouldRefreshOnlyPerBrokerConfigs(testCase.CurrentConfigs, testCase.DesiredConfigs, logger) != testCase.Result {
+			t.Errorf("test case %d failed: %s", i, testCase.Description)
+		}
+	}
+}
+
+func TestCollectTouchedConfigs(t *testing.T) {
+	testCases := []struct {
+		DesiredConfigs string
+		CurrentConfigs string
+		Result         []string
+	}{
+		{
+			DesiredConfigs: `key1=value1
+key2=value2
+key3=value3`,
+			CurrentConfigs: `key1=value1
+key2=value2
+key3=value3`,
+			Result: []string{},
+		},
+		{
+			DesiredConfigs: `key1=value1
+key2=value2`,
+			CurrentConfigs: `key1=value1
+key2=value2
+key3=value3`,
+			Result: []string{"key3"},
+		},
+		{
+			DesiredConfigs: `key1=value1
+key2=value2
+key3=value3`,
+			CurrentConfigs: `key1=value1
+key2=value2`,
+			Result: []string{"key3"},
+		},
+		{
+			DesiredConfigs: `key1=value1
+key2=value2
+key3=value3`,
+			CurrentConfigs: `key1=value1
+key2=value4
+key3=value3`,
+			Result: []string{"key2"},
+		},
+		{
+			DesiredConfigs: `key1=value1
+key2=value2
+key3=value3
+key4=value4`,
+			CurrentConfigs: `key1=value6
+key2=value2
+key3=value3
+key5=value5`,
+			Result: []string{"key1", "key4", "key5"},
+		},
+	}
+
+	logger := util.CreateLogger(false, false)
+	for _, testCase := range testCases {
+		desired := &corev1.ConfigMap{
+			Data: map[string]string{
+				"broker-config": testCase.DesiredConfigs,
+			},
+		}
+		current := &corev1.ConfigMap{
+			Data: map[string]string{
+				"broker-config": testCase.CurrentConfigs,
+			},
+		}
+		touchedConfigs := collectTouchedConfigs(getBrokerConfigsFromConfigMap(current), getBrokerConfigsFromConfigMap(desired), logger)
+		sort.Strings(touchedConfigs)
+		if !reflect.DeepEqual(touchedConfigs, testCase.Result) {
+			t.Errorf("comparison failed - expected: %s, actual: %s", testCase.Result, touchedConfigs)
+		}
 	}
 }
