+++ pkg/resources/kafka/configs.go
@@ -15,24 +15,21 @@
 package kafka
 
 import (
-	"context"
-	"reflect"
 	"strconv"
 	"strings"
 
 	"emperror.dev/errors"
 	"github.com/Shopify/sarama"
 	"github.com/go-logr/logr"
-
 	corev1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/banzaicloud/kafka-operator/api/v1beta1"
 	"github.com/banzaicloud/kafka-operator/pkg/errorfactory"
 	"github.com/banzaicloud/kafka-operator/pkg/k8sutil"
 	"github.com/banzaicloud/kafka-operator/pkg/util"
 )
+
+const securityProtocolMap = "listener.security.protocol.map"
 
 // these configurations will not trigger rolling upgrade when updated
 var perBrokerConfigs = []string{
@@ -42,9 +39,9 @@ var perBrokerConfigs = []string{
 	// listener related config change will trigger rolling upgrade anyways due to pod spec change
 	"listeners",
 	"advertised.listeners",
-}
 
-const securityProtocolMap = "listener.security.protocol.map"
+	securityProtocolMap,
+}
 
 func (r *Reconciler) reconcilePerBrokerDynamicConfig(brokerId int32, brokerConfig *v1beta1.BrokerConfig, configMap *corev1.ConfigMap, log logr.Logger) error {
 	kClient, err := r.kafkaClientProvider.NewFromCluster(r.Client, r.KafkaCluster)
@@ -57,35 +54,32 @@ func (r *Reconciler) reconcilePerBrokerDynamicConfig(brokerId int32, brokerConfi
 		}
 	}()
 
-	parsedBrokerConfig := util.ParsePropertiesFormat(brokerConfig.Config)
+	fullPerBrokerConfig := util.ParsePropertiesFormat(brokerConfig.Config)
 
 	currentPerBrokerConfigState := r.KafkaCluster.Status.BrokersState[strconv.Itoa(int(brokerId))].PerBrokerConfigurationState
-	if len(parsedBrokerConfig) == 0 && currentPerBrokerConfigState != v1beta1.PerBrokerConfigOutOfSync {
+	if len(fullPerBrokerConfig) == 0 && currentPerBrokerConfigState != v1beta1.PerBrokerConfigOutOfSync {
 		return nil
 	}
 
-	// overwrite configs that are generated in the configmap
+	// overwrite configs from configmap
 	configsFromConfigMap := getBrokerConfigsFromConfigMap(configMap)
 	for _, perBrokerConfig := range perBrokerConfigs {
 		if configValue, ok := configsFromConfigMap[perBrokerConfig]; ok {
-			parsedBrokerConfig[perBrokerConfig] = configValue
+			fullPerBrokerConfig[perBrokerConfig] = configValue
 		}
 	}
 
-	brokerConfigKeys := make([]string, 0, len(parsedBrokerConfig)+len(perBrokerConfigs))
-	for key := range parsedBrokerConfig {
+	// query the current configs
+	brokerConfigKeys := make([]string, 0, len(fullPerBrokerConfig))
+	for key := range fullPerBrokerConfig {
 		brokerConfigKeys = append(brokerConfigKeys, key)
 	}
-	brokerConfigKeys = append(brokerConfigKeys, perBrokerConfigs...)
-
 	response, err := kClient.DescribePerBrokerConfig(brokerId, brokerConfigKeys)
 	if err != nil {
 		return errors.WrapIfWithDetails(err, "could not describe broker config", "brokerId", brokerId)
 	}
 
-	changedConfigs := collectNonIdenticalConfigsFromResponse(response, parsedBrokerConfig)
-
-	if len(changedConfigs) > 0 {
+	if shouldUpdatePerBrokerConfig(response, fullPerBrokerConfig) {
 		if currentPerBrokerConfigState == v1beta1.PerBrokerConfigInSync {
 			log.V(1).Info("setting per broker config status to out of sync")
 			statusErr := k8sutil.UpdateBrokerStatus(r.Client, []string{strconv.Itoa(int(brokerId))}, r.KafkaCluster, v1beta1.PerBrokerConfigOutOfSync, log)
@@ -95,7 +89,7 @@ func (r *Reconciler) reconcilePerBrokerDynamicConfig(brokerId int32, brokerConfi
 		}
 
 		// validate the config
-		err := kClient.AlterPerBrokerConfig(brokerId, util.ConvertMapStringToMapStringPointer(changedConfigs), true)
+		err := kClient.AlterPerBrokerConfig(brokerId, util.ConvertMapStringToMapStringPointer(fullPerBrokerConfig), true)
 		if err != nil {
 			statusErr := k8sutil.UpdateBrokerStatus(r.Client, []string{strconv.Itoa(int(brokerId))}, r.KafkaCluster, v1beta1.PerBrokerConfigError, log)
 			if statusErr != nil {
@@ -105,20 +99,19 @@ func (r *Reconciler) reconcilePerBrokerDynamicConfig(brokerId int32, brokerConfi
 		}
 
 		// alter the config
-		err = kClient.AlterPerBrokerConfig(brokerId, util.ConvertMapStringToMapStringPointer(changedConfigs), false)
+		err = kClient.AlterPerBrokerConfig(brokerId, util.ConvertMapStringToMapStringPointer(fullPerBrokerConfig), false)
 		if err != nil {
 			return errors.WrapIfWithDetails(err, "could not alter broker config", "brokerId", brokerId)
 		}
 
-		// request for the updated config
+		// query the updated config
 		response, err := kClient.DescribePerBrokerConfig(brokerId, brokerConfigKeys)
 		if err != nil {
 			return errors.WrapIfWithDetails(err, "could not describe broker config", "brokerId", brokerId)
 		}
 
-		// update per-broker config status
-		notUpdatedConfigs := collectNonIdenticalConfigsFromResponse(response, parsedBrokerConfig)
-		if len(notUpdatedConfigs) > 0 {
+		// update per-broker config status based on the response
+		if shouldUpdatePerBrokerConfig(response, fullPerBrokerConfig) {
 			return errorfactory.New(errorfactory.PerBrokerConfigNotReady{}, errors.New("configuration is out of sync"), "per-broker configuration updated")
 		} else {
 			statusErr := k8sutil.UpdateBrokerStatus(r.Client, []string{strconv.Itoa(int(brokerId))}, r.KafkaCluster, v1beta1.PerBrokerConfigInSync, log)
@@ -184,71 +177,6 @@ func (r *Reconciler) reconcileClusterWideDynamicConfig(log logr.Logger) error {
 	return nil
 }
 
-func (r *Reconciler) postConfigMapReconcile(log logr.Logger, desired *corev1.ConfigMap) error {
-	desiredType := reflect.TypeOf(desired)
-	current := desired.DeepCopyObject()
-
-	key, err := runtimeClient.ObjectKeyFromObject(current)
-	if err != nil {
-		return errors.WithDetails(err, "kind", desiredType)
-	}
-	log = log.WithValues("kind", desiredType, "name", key.Name)
-
-	err = r.Client.Get(context.TODO(), key, current)
-	if err != nil && !apierrors.IsNotFound(err) {
-		return errorfactory.New(errorfactory.APIFailure{}, err, "getting resource failed", "kind", desiredType, "name", key.Name)
-	}
-
-	if k8sutil.CheckIfObjectUpdated(log, desiredType, current, desired) {
-		// Only update status when configmap belongs to broker
-		if id, ok := desired.Labels["brokerId"]; ok {
-			currentConfigs := getBrokerConfigsFromConfigMap(current.(*corev1.ConfigMap))
-			desiredConfigs := getBrokerConfigsFromConfigMap(desired)
-
-			var statusErr error
-			// if only per broker configs are changed, do not trigger rolling upgrade by setting ConfigOutOfSync status
-			if shouldRefreshOnlyPerBrokerConfigs(currentConfigs, desiredConfigs, log) {
-				log.V(1).Info("setting per broker config status to out of sync")
-				statusErr = k8sutil.UpdateBrokerStatus(r.Client, []string{id}, r.KafkaCluster, v1beta1.PerBrokerConfigOutOfSync, log)
-			} else {
-				statusErr = k8sutil.UpdateBrokerStatus(r.Client, []string{id}, r.KafkaCluster, v1beta1.ConfigOutOfSync, log)
-			}
-			if statusErr != nil {
-				return errors.WrapIfWithDetails(err, "updating status for resource failed", "kind", desiredType)
-			}
-		}
-	}
-
-	return nil
-}
-
-// collects are the config keys that are either added, removed or updated
-// between the current and the desired ConfigMap
-func collectTouchedConfigs(currentConfigs, desiredConfigs map[string]string, log logr.Logger) []string {
-	touchedConfigs := make([]string, 0)
-
-	currentConfigsCopy := make(map[string]string, 0)
-	for k, v := range currentConfigs {
-		currentConfigsCopy[k] = v
-	}
-
-	for configName, desiredValue := range desiredConfigs {
-		if currentValue, ok := currentConfigsCopy[configName]; !ok || currentValue != desiredValue {
-			// new or updated config
-			touchedConfigs = append(touchedConfigs, configName)
-		}
-		delete(currentConfigsCopy, configName)
-	}
-
-	for configName := range currentConfigsCopy {
-		// deleted config
-		touchedConfigs = append(touchedConfigs, configName)
-	}
-
-	log.V(1).Info("configs have been changed", "configs", touchedConfigs)
-	return touchedConfigs
-}
-
 func getBrokerConfigsFromConfigMap(configMap *corev1.ConfigMap) map[string]string {
 	brokerConfig := configMap.Data["broker-config"]
 	configs := strings.Split(brokerConfig, "\n")
@@ -263,64 +191,14 @@ func getBrokerConfigsFromConfigMap(configMap *corev1.ConfigMap) map[string]strin
 	return m
 }
 
-func collectNonIdenticalConfigsFromResponse(response []*sarama.ConfigEntry, parsedBrokerConfig map[string]string) map[string]string {
-	changedConfigs := make(map[string]string, 0)
+func shouldUpdatePerBrokerConfig(response []*sarama.ConfigEntry, parsedBrokerConfig map[string]string) bool {
 	for _, conf := range response {
 		if val, ok := parsedBrokerConfig[conf.Name]; ok {
 			if val != conf.Value {
-				changedConfigs[conf.Name] = val
+				return true
 			}
 		}
 	}
 
-	return changedConfigs
-}
-
-func shouldRefreshOnlyPerBrokerConfigs(currentConfigs, desiredConfigs map[string]string, log logr.Logger) bool {
-	touchedConfigs := collectTouchedConfigs(currentConfigs, desiredConfigs, log)
-
-OUTERLOOP:
-	for _, config := range touchedConfigs {
-		for _, perBrokerConfig := range perBrokerConfigs {
-			if config == perBrokerConfig {
-				continue OUTERLOOP
-			}
-		}
-
-		// Security protocol cannot be updated for existing listener
-		// a rolling upgrade should be triggered in this case
-		if config == securityProtocolMap {
-			// added or deleted config is ok
-			if currentConfigs[securityProtocolMap] == "" || desiredConfigs[securityProtocolMap] == "" {
-				continue
-			}
-			currentListenerProtocolMap := make(map[string]string, 0)
-			desiredListenerProtocolMap := make(map[string]string, 0)
-			for _, listenerConfig := range strings.Split(currentConfigs[securityProtocolMap], ",") {
-				listenerKeyValue := strings.Split(listenerConfig, ":")
-				if len(listenerKeyValue) != 2 {
-					continue
-				}
-				currentListenerProtocolMap[listenerKeyValue[0]] = listenerKeyValue[1]
-			}
-			for _, listenerConfig := range strings.Split(desiredConfigs[securityProtocolMap], ",") {
-				listenerKeyValue := strings.Split(listenerConfig, ":")
-				if len(listenerKeyValue) != 2 {
-					continue
-				}
-				desiredListenerProtocolMap[listenerKeyValue[0]] = listenerKeyValue[1]
-			}
-
-			for protocolName, desiredListenerProtocol := range desiredListenerProtocolMap {
-				if currentListenerProtocol, ok := currentListenerProtocolMap[protocolName]; ok {
-					if currentListenerProtocol != desiredListenerProtocol {
-						return false
-					}
-				}
-			}
-			continue
-		}
-		return false
-	}
-	return true
+	return false
 }
