+++ pkg/resources/kafka/configs_test.go
@@ -16,14 +16,84 @@ package kafka
 
 import (
 	"reflect"
-	"sort"
 	"testing"
 
 	"github.com/Shopify/sarama"
 	corev1 "k8s.io/api/core/v1"
-
-	"github.com/banzaicloud/kafka-operator/pkg/util"
 )
+
+func TestCollectNonIdenticalConfigsFromResponse(t *testing.T) {
+	testCases := []struct {
+		Description  string
+		Response     []*sarama.ConfigEntry
+		BrokerConfig map[string]string
+		Result       bool
+	}{
+		{
+			Description: "matching configs",
+			Response: []*sarama.ConfigEntry{
+				{
+					Name:  "config1",
+					Value: "value1",
+				},
+				{
+					Name:  "config2",
+					Value: "value2",
+				},
+			},
+			BrokerConfig: map[string]string{
+				"config1": "value1",
+				"config2": "value2",
+			},
+			Result: false,
+		},
+		{
+			Description: "if a config is deleted, do not update the broker",
+			Response: []*sarama.ConfigEntry{
+				{
+					Name:  "config1",
+					Value: "value1",
+				},
+				{
+					Name:  "config2",
+					Value: "value2",
+				},
+				{
+					Name:  "config3",
+					Value: "value3",
+				},
+			},
+			BrokerConfig: map[string]string{
+				"config1": "value1",
+				"config2": "value2",
+			},
+			Result: false,
+		},
+		{
+			Description: "config value updated",
+			Response: []*sarama.ConfigEntry{
+				{
+					Name:  "config1",
+					Value: "value1",
+				},
+				{
+					Name:  "config2",
+					Value: "value2",
+				},
+			},
+			BrokerConfig: map[string]string{
+				"config1": "value1",
+				"config2": "value3",
+			},
+			Result: true,
+		},
+	}
+	for _, testCase := range testCases {
+		if shouldUpdatePerBrokerConfig(testCase.Response, testCase.BrokerConfig) != testCase.Result {
+			t.Errorf("wrong result generated by test case: %s", testCase.Description)
+		}
+	}
+}
 
 func TestGetBrokerConfigsFromConfigMap(t *testing.T) {
 	configMap := &corev1.ConfigMap{
@@ -41,246 +111,5 @@ key3=value3`,
 	actual := getBrokerConfigsFromConfigMap(configMap)
 	if !reflect.DeepEqual(expected, actual) {
 		t.Errorf("broker config extraction from configmap failed - expected: %s, actual: %s", expected, actual)
-	}
-}
-
-func TestCollectTouchedConfigs(t *testing.T) {
-	testCases := []struct {
-		DesiredConfigs string
-		CurrentConfigs string
-		Result         []string
-	}{
-		{
-			DesiredConfigs: `key1=value1
-key2=value2
-key3=value3`,
-			CurrentConfigs: `key1=value1
-key2=value2
-key3=value3`,
-			Result: []string{},
-		},
-		{
-			DesiredConfigs: `key1=value1
-key2=value2`,
-			CurrentConfigs: `key1=value1
-key2=value2
-key3=value3`,
-			Result: []string{"key3"},
-		},
-		{
-			DesiredConfigs: `key1=value1
-key2=value2
-key3=value3`,
-			CurrentConfigs: `key1=value1
-key2=value2`,
-			Result: []string{"key3"},
-		},
-		{
-			DesiredConfigs: `key1=value1
-key2=value2
-key3=value3`,
-			CurrentConfigs: `key1=value1
-key2=value4
-key3=value3`,
-			Result: []string{"key2"},
-		},
-		{
-			DesiredConfigs: `key1=value1
-key2=value2
-key3=value3
-key4=value4`,
-			CurrentConfigs: `key1=value6
-key2=value2
-key3=value3
-key5=value5`,
-			Result: []string{"key1", "key4", "key5"},
-		},
-	}
-
-	logger := util.CreateLogger(false, false)
-	for _, testCase := range testCases {
-		desired := &corev1.ConfigMap{
-			Data: map[string]string{
-				"broker-config": testCase.DesiredConfigs,
-			},
-		}
-		current := &corev1.ConfigMap{
-			Data: map[string]string{
-				"broker-config": testCase.CurrentConfigs,
-			},
-		}
-		touchedConfigs := collectTouchedConfigs(getBrokerConfigsFromConfigMap(current), getBrokerConfigsFromConfigMap(desired), logger)
-		sort.Strings(touchedConfigs)
-		if !reflect.DeepEqual(touchedConfigs, testCase.Result) {
-			t.Errorf("comparison failed - expected: %s, actual: %s", testCase.Result, touchedConfigs)
-		}
-	}
-}
-
-func TestCollectNonIdenticalConfigsFromResponse(t *testing.T) {
-	testCases := []struct {
-		Description  string
-		Response     []*sarama.ConfigEntry
-		BrokerConfig map[string]string
-		Result       map[string]string
-	}{
-		{
-			Description: "matching configs",
-			Response: []*sarama.ConfigEntry{
-				{
-					Name: "config1",
-					Value: "value1",
-				},
-				{
-					Name: "config2",
-					Value: "value2",
-				},
-			},
-			BrokerConfig: map[string]string{
-				"config1": "value1",
-				"config2": "value2",
-			},
-			Result: map[string]string{},
-		},
-		{
-			Description: "if a config is deleted, do not update the broker",
-			Response: []*sarama.ConfigEntry{
-				{
-					Name: "config1",
-					Value: "value1",
-				},
-				{
-					Name: "config2",
-					Value: "value2",
-				},
-				{
-					Name: "config3",
-					Value: "value3",
-				},
-			},
-			BrokerConfig: map[string]string{
-				"config1": "value1",
-				"config2": "value2",
-			},
-			Result: map[string]string{},
-		},
-		{
-			Description: "config value updated",
-			Response: []*sarama.ConfigEntry{
-				{
-					Name: "config1",
-					Value: "value1",
-				},
-				{
-					Name: "config2",
-					Value: "value2",
-				},
-			},
-			BrokerConfig: map[string]string{
-				"config1": "value1",
-				"config2": "value3",
-			},
-			Result: map[string]string{
-				"config2": "value3",
-			},
-		},
-	}
-	for _, testCase := range testCases {
-		actualResult := collectNonIdenticalConfigsFromResponse(testCase.Response, testCase.BrokerConfig)
-		if !reflect.DeepEqual(testCase.Result, actualResult) {
-			t.Errorf("wrong result generated by test case %s - expected: %s, actual: %s", testCase.Description, testCase.Result, actualResult)
-		}
-	}
-}
-
-func TestShouldRefreshOnlyPerBrokerConfigs(t *testing.T) {
-	testCases := []struct {
-		Description    string
-		CurrentConfigs map[string]string
-		DesiredConfigs map[string]string
-		Result         bool
-	}{
-		{
-			Description: "configs did not change",
-			CurrentConfigs: map[string]string{
-				"unmodified_config_1": "unmodified_value_1",
-				"unmodified_config_2": "unmodified_value_2",
-			},
-			DesiredConfigs: map[string]string{
-				"unmodified_config_1": "unmodified_value_1",
-				"unmodified_config_2": "unmodified_value_2",
-			},
-			Result: true,
-		},
-		{
-			Description: "only non per-broker config changed",
-			CurrentConfigs: map[string]string{
-				"unmodified_config_1": "unmodified_value_1",
-				"unmodified_config_2": "modified_value_2",
-			},
-			DesiredConfigs: map[string]string{
-				"unmodified_config_1": "unmodified_value_1",
-				"unmodified_config_2": "modified_value_3",
-			},
-			Result: false,
-		},
-		{
-			Description: "only per-broker config changed",
-			CurrentConfigs: map[string]string{
-				"unmodified_config_1": "unmodified_value_1",
-				"ssl.client.auth":     "modified_value_2",
-			},
-			DesiredConfigs: map[string]string{
-				"unmodified_config_1": "unmodified_value_1",
-				"ssl.client.auth":     "modified_value_3",
-			},
-			Result: true,
-		},
-		{
-			Description: "per-broker and non per-broker configs have changed",
-			CurrentConfigs: map[string]string{
-				"modified_config_1": "modified_value_1",
-				"ssl.client.auth":   "modified_value_3",
-			},
-			DesiredConfigs: map[string]string{
-				"modified_config_1": "modified_value_2",
-				"ssl.client.auth":   "modified_value_4",
-			},
-			Result: false,
-		},
-		{
-			Description: "security protocol map can be changed as a per-broker config",
-			CurrentConfigs: map[string]string{
-				"listener.security.protocol.map": "listener1:protocol1,listener2:protocol2",
-			},
-			DesiredConfigs: map[string]string{
-				"listener.security.protocol.map": "listener1:protocol1,listener3:protocol3",
-			},
-			Result: true,
-		},
-		{
-			Description: "security protocol map can't be changed as a per-broker config",
-			CurrentConfigs: map[string]string{
-				"listener.security.protocol.map": "listener1:protocol1,listener2:protocol2",
-			},
-			DesiredConfigs: map[string]string{
-				"listener.security.protocol.map": "listener1:protocol1,listener2:protocol3",
-			},
-			Result: false,
-		},
-		{
-			Description: "security protocol map added as config",
-			CurrentConfigs: map[string]string{},
-			DesiredConfigs: map[string]string{
-				"listener.security.protocol.map": "listener1:protocol1,listener2:protocol2",
-			},
-			Result: true,
-		},
-	}
-	logger := util.CreateLogger(false, false)
-	for i, testCase := range testCases {
-		if shouldRefreshOnlyPerBrokerConfigs(testCase.CurrentConfigs, testCase.DesiredConfigs, logger) != testCase.Result {
-			t.Errorf("test case %d failed: %s", i, testCase.Description)
-		}
 	}
 }
